+++ client/src/helpers/auth.helpers.ts
